port modifyGroup function to use jsonrpc
