+++ src/main/deltachat/controller.ts
@@ -3,7 +3,6 @@ import { app as rawApp, ipcMain } from 'electron'
 import { EventEmitter } from 'events'
 import { getLogger } from '../../shared/logger'
 import * as mainWindow from '../windows/main'
-import DCChat from './chat'
 import DCContext from './context'
 import DCLoginController from './login'
 import DCMessageList from './messagelist'
@@ -233,7 +232,6 @@ export default class DeltaChatController extends EventEmitter {
     logMigrate.info('migration completed')
   }
 
-  readonly chat = new DCChat(this)
   readonly login = new DCLoginController(this)
   readonly messageList = new DCMessageList(this)
   readonly stickers = new DCStickers(this)
